refactor: move api hooks into generic hooks directory
