+++ src/components/image/AccessibilityCloudImage.tsx
@@ -2,8 +2,8 @@
 
 import NextImage, { type ImageProps } from "next/image";
 import { useCallback, useMemo } from "react";
+import useAccessibilityCloudApi from "~/hooks/useAccessibilityCloudApi";
 import type { AccessibilityCloudImage as AccessibilityCloudImageType } from "~/needs-refactoring/lib/model/ac/Feature";
-import useAccessibilityCloud from "~/modules/accessibility-cloud/hooks/useAccessibilityCloud";
 
 const maxImageSize = 1920;
 const defaultPlaceholder =
@@ -63,7 +63,7 @@ export default function AccessibilityCloudImage({
   image: AccessibilityCloudImageType;
   forceReload?: boolean; // Useful for testing. Will force a reload of the image from the server on every rerender
 }) {
-  const { baseUrl } = useAccessibilityCloud({ cached: true });
+  const { baseUrl } = useAccessibilityCloudApi({ cached: true });
 
   const additionalQueryParameter = forceReload
     ? `version=${Math.random().toString().slice(2)}`
