+++ src/needs-refactoring/lib/fetchers/useFeatures/useOsmToAcFeature.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import useSWRInfinite, { type SWRInfiniteConfiguration } from "swr/infinite";
-import useAccessibilityCloud from "~/modules/accessibility-cloud/hooks/useAccessibilityCloud";
+import useAccessibilityCloudApi from "~/hooks/useAccessibilityCloudApi";
 import type { OSMIdWithTableAndContextNames } from "../../typing/brands/osmIds";
 import { getOSMRDFComponents } from "../../typing/discriminators/osmDiscriminator";
 import {
@@ -19,9 +19,11 @@ export const useOsmToAcFeature = (
     cache?: boolean;
   },
 ) => {
-  const { baseUrl: acBaseUrl, appToken: acAppToken } = useAccessibilityCloud({
-    cached: Boolean(options?.cache),
-  });
+  const { baseUrl: acBaseUrl, appToken: acAppToken } = useAccessibilityCloudApi(
+    {
+      cached: Boolean(options?.cache),
+    },
+  );
 
   const [properties, uris] = useMemo(() => {
     // building additional properties, such as the fetcher can add additional attributes
